Refactor Signup into separate file
