+++ www/src/App.tsx
@@ -21,7 +21,7 @@ const Plural = lazy(() => import('./components/Plural'))
 const Invite = lazy(() => import('./components/Invite'))
 const Login = lazy(() => import('./components/users/MagicLogin').then(module => ({ default: module.Login })))
 const PasswordlessLogin = lazy(() => import('./components/users/MagicLogin').then(module => ({ default: module.PasswordlessLogin })))
-const Signup = lazy(() => import('./components/users/MagicLogin').then(module => ({ default: module.Signup })))
+const Signup = lazy(() => import('./components/users/Signup').then(module => ({ default: module.Signup })))
 const PasswordReset = lazy(() => import('./components/users/PasswordReset').then(module => ({ default: module.PasswordReset })))
 const ResetPassword = lazy(() => import('./components/users/PasswordReset').then(module => ({ default: module.ResetPassword })))
 const OAuthConsent = lazy(() => import('./components/oidc/OAuthConsent').then(module => ({ default: module.OAuthConsent })))
